pi/requirements-manual.txt
pyserial>=3.5
evdev>=1.7.1
